MariTalk/requirements.txt
maritalk==0.2.0
python-dotenv==1.0.1